refactor: remove redundant files and reorganize test files
