+++ packages/validators/test/__fixtures__/v2023_09_28/experiencePattern.ts
